mcp-server/requirements.txt
azure-functions
azure-storage-blob
aiohttp
pandas
